# Remove NewPossibleNonCounterInfo until it can be made more efficient