According to benchmarks, this is very inefficient for many rate queries, temporarily removing until we find a better way.

With before as without the removed code, and after as with:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                                                  │ BenchmarkRangeQuery_before.txt │    BenchmarkRangeQuery_after.txt     │
                                                  │             sec/op             │   sec/op     vs base                 │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                      15.33m ± 2%   89.57m ± 3%  +484.36% (p=0.000 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │      BenchmarkRangeQuery_after.txt      │
                                                  │              B/op              │     B/op       vs base                  │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                     4.946Mi ± 0%   66.805Mi ± 0%  +1250.79% (p=0.000 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │    BenchmarkRangeQuery_after.txt     │
                                                  │           allocs/op            │  allocs/op   vs base                 │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                      105.4k ± 0%   806.3k ± 0%  +665.22% (p=0.000 n=10)
```