@bboreham 2023-10-20
I compared commit 156222cc5, before annotations were added, with this PR, commit 7756556eb:

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                                 │   before    │               after                │
                                                 │   sec/op    │   sec/op     vs base               │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   29.82m ± 3%   38.52m ± 3%  +29.19% (p=0.002 n=6)

                                                 │    before    │                 after                  │
                                                 │     B/op     │     B/op       vs base                 │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   378.8Ki ± 0%   5063.3Ki ± 0%  +1236.53% (p=0.002 n=6)

                                                 │   before    │                 after                  │
                                                 │  allocs/op  │   allocs/op    vs base                 │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   5.356k ± 0%   105.365k ± 0%  +1867.24% (p=0.002 n=6)
```

12x more garbage does not seem acceptable to me, so I pushed another commit which eliminates it in the case there is nothing to report.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                                 │   before    │               after               │
                                                 │   sec/op    │   sec/op     vs base              │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   29.82m ± 3%   31.06m ± 1%  +4.17% (p=0.002 n=6)

                                                 │    before    │               after                │
                                                 │     B/op     │     B/op      vs base              │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   378.8Ki ± 0%   368.1Ki ± 0%  -2.84% (p=0.002 n=6)

                                                 │   before    │               after               │
                                                 │  allocs/op  │  allocs/op   vs base              │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4   5.356k ± 0%   5.258k ± 0%  -1.83% (p=0.002 n=6)
```

The 4.17% extra CPU comes from `Metric.Get(labels.MetricName)`, which can also be eliminated, but I wanted to do the minimum to get to a reasonable state.