@zenador 2023-10-20
Thank you! I added a commit to further avoid creating empty annotations, and created a PR with the same changes for the new prometheus release (benchmark details there): https://github.com/grafana/mimir-prometheus/pull/547